containers:  add metavar for rebase option
The inline help for `atomic containers update` currently shows
`--rebase=REBASE` which is not very helpful in terms of what kind of
argument is supposed to be passed to that flag.

This change just adds a `metavar` to show that users should pass in an
image name to the `--rebase` flag.